+++ src/backbone.py
@@ -30,16 +30,18 @@ class Backbone:
             Returns:
                 model (tf.keras.Model): The ResNet50 model.
         '''
+
+        input_image = tf.keras.layers.Input(shape=self.config.image_shape, batch_size=self.config.batch_size, name='input_image')
         
         # Load a pre-trained ResNet50 model
-        resnet50 = tf.keras.applications.ResNet50(weights=self.config.backbone_weights,
-                                                  include_top=False,
-                                                  input_shape=self.config.image_shape)
+        resnet50 = tf.keras.applications.ResNet50(input_tensor=input_image,
+                                                  weights=self.config.backbone_weights,
+                                                  include_top=False)
 
         # Freeze all layers except the last `trainable_layers`
         for layer in resnet50.layers[:-self.config.backbone_trainable_layers]:
             layer.trainable = False
 
-        model = tf.keras.Model(inputs=resnet50.inputs, outputs=resnet50.output)
+        model = tf.keras.Model(inputs=input_image, outputs=resnet50.output)
         return model