+++ src/utils.py
@@ -31,7 +31,7 @@ class Config:
         # General
         self.image_shape = (250, 250, 3)
         self.num_classes = 2
-        self.batch_size = 32
+        self.batch_size = 16
         self.model_dir = '../model/'
 
         # Utils